✨ feat(codecs): tuple2 record helpers + int-typed codecs
- solana_kit_codecs_data_structures: typed getTuple2Encoder/Decoder for
  two-element tuples exposed as Dart records.
- solana_kit_codecs_numbers: integer encoders (u8/i8/u16/i16/u32) narrowed
  to FixedSizeEncoder<int> via a generic numberEncoderFactory so generated
  strict contexts compile.
- codama-renderers-dart: emit getTuple2* for arity-2 tuple nodes, escape
  Dart reserved-word identifiers, keep generated instructionData locals
  collision-free.
- Both codec packages are at 100% line coverage; changesets added.

Researched and implemented on behalf of Ifiok Jr. (@ifiokjr) by pi (codex, high thinking).

+++ packages/solana_kit_codecs_numbers/lib/src/i16.dart
@@ -12,7 +12,7 @@ FixedSizeEncoder<int> getI16Encoder([NumberCodecConfig? config]) =>
       name: 'i16',
       size: 2,
       set: (data, offset, value, endian) =>
-          data.setInt16(offset, value.toInt(), endian),
+          data.setInt16(offset, value, endian),
       config: config,
       range: (-32768, 32767),
     );
